% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_limma.R
\name{diff_limma}
\alias{diff_limma}
\title{Sample-level differential analysis}
\usage{
diff_limma(
  feature_list,
  feature_type,
  hierarchy_list,
  sample_meta,
  response_variable,
  separate = T,
  differential_method = "limma",
  formula = NULL,
  coef = 2,
  fdr_cutoff = 0.05,
  column_name = "sample",
  ncores = parallel::detectCores(),
  save_as_csv = T,
  verbose = T
)
}
\arguments{
\item{feature_list}{A list of sample-level feature matrices. Default is NULL.}

\item{feature_type}{Name of input feature type, will be used as suffix of output file and column names.}

\item{hierarchy_list}{a hierarchy list by running \code{'extract_hrchy_string()'} or \code{'extract_hrchy_seurat()'} function.}

\item{sample_meta}{Sample-level metadata, where each row is a sample. Must contain 'sample' column and additional variables to be used in the analysis, such as covariates or outcomes of interest.}

\item{response_variable}{A vector of response variables.}

\item{separate}{A TRUE (default) or FALSE indicator, specifying how to evaluate multivariate outcomes.
\itemize{
\item TRUE: evaluate multivariate phenotype separately (it is equivalent to run this pipeline for each univariate phenotype).
\item FALSE: evaluate multivariate phenotype jointly.
}}

\item{differential_method}{Specify which type of differential methods: limma (default) or gam}

\item{formula}{An object of class 'formula': a symbolic description of adjustment formula (i.e. only includes covariates other than response variable(s))}

\item{coef}{A column number or column name specifying which coefficient to be extracted (by default: 2).}

\item{fdr_cutoff}{Cutoff value for FDR. Only genes with lower FDR are listed. Default is 0.05.}

\item{column_name}{column name of feature matrix. default is 'sample.}

\item{ncores}{Number of cores to be used. If ncores > 1, it will be implemented in a parallel mode.}

\item{save_as_csv}{An indicator to save identified DEGs in csv files. DEGs for each cell cluster is saved as 'responsevariable_celltype_DEG.csv' and a summary file of combining DEGs from all cell clusters is saved as 'responsevariable_combinedDEG.csv'.}

\item{verbose}{Show progress}
}
\value{
A list of three elements:
\itemize{
\item diff.summary: A summary table of number of differential features for each tree node.
\item diff.ls: A comprehensive list of outcome(s)-associated feature set for each tree node. Use \code{`result$dge.ls$response_variable[[celltype]]`} to extract DEGs for a specific cell type
\item feature.ls: A list of input feature list
\item diff.summary.bygene: A summary table for each differentiated gene.
}
}
\description{
Provides feature sets and uses LIMMA to identify differentially features for each cell cluster.
}
\author{
Boyang Zhang <bzhang34@jhu.edu>, Hongkai Ji
}
