% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContactRegion.R
\name{getContactRegion}
\alias{getContactRegion}
\title{Get Contact Region (remove residues)}
\usage{
getContactRegion(.data, .start = 4, .end = -4)
}
\arguments{
\item{.data}{data frame}

\item{.start}{starting position}

\item{.end}{ending position}
}
\description{
Get Contact Region (remove residues)
}
