% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExpandedClone.R
\name{getExpandedClone}
\alias{getExpandedClone}
\title{Identify exapnded clones}
\usage{
getExpandedClone(
  data,
  match = "clonotype",
  type = "count",
  threshold = 3,
  aggregate = "sample",
  ncores = parallel::detectCores()
)
}
\arguments{
\item{data}{a list of immune profiling info for each sample. in `immunarch` format. It corresponds to extract [['data']] slot in `immunarch`.}

\item{match}{match by 'clonotype' or 'clonotype+vdj'}

\item{type}{use 'count' or 'frequency' to define expanded clones}

\item{threshold}{specify a minimum threshold for identifying expanded clones for each sample. Default threshold is 3. Clones shared among at least `threshold` number of cells would be identified as expanded clones.}

\item{aggregate}{specify either aggregate at sample-level ('sample') or combine all samples together ('all')}

\item{ncores}{number of cores}
}
\description{
Identify exapnded clones
}
\author{
Boyang Zhang <bzhang34@jhu.edu>
}
