% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneUsage.R
\name{getGeneUsage}
\alias{getGeneUsage}
\title{Get gene usage}
\usage{
getGeneUsage(
  data,
  gene,
  norm = T,
  transform = "none",
  ncores = parallel::detectCores(),
  filter_prop = 0.1
)
}
\arguments{
\item{data}{a list of immune profiling info for each sample. in `immunarch` format. It corresponds to extract [['data']] slot in `immunarch`.}

\item{gene}{specify one or multiple queried tcr or bcr genes (use '+' to separate).}

\item{norm}{an indicator to specify normalization or not. Default is TRUE (normalized to frequency).}

\item{transform}{specify either 'clr' (centered log ratio) transformation or 'none' (default).}

\item{ncores}{number of cores}
}
\description{
Get gene usage
}
\author{
Boyang Zhang <bzhang34@jhu.edu>
}
