% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUnqKmers.R
\name{getUnqKmers}
\alias{getUnqKmers}
\title{get unique kmer per sequence}
\usage{
getUnqKmers(
  .data,
  .k,
  .col = c("aa", "nt"),
  .coding = TRUE,
  ncores = parallel::detectCores()
)
}
\arguments{
\item{.data}{immunarch data}

\item{.k}{specify kmer length}

\item{.col}{specify amino acid or nucleotide}

\item{.coding}{coding region}

\item{ncores}{number of cores}
}
\description{
get unique kmer per sequence
}
