% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{plot_heatmap}
\usage{
plot_heatmap(
  mat,
  sample_meta,
  scale = "row",
  sort_bycol = NA,
  cluster_rows = F,
  cluster_cols = F,
  show_colnames = F,
  show_rownames = T,
  annotation_col = NA,
  annotation_colors = NA,
  color = colorRampPalette(rev(brewer.pal(n = 7, name = "RdYlBu")))(100),
  breaks = NA,
  fontsize = 12
)
}
\description{
plot_heatmap
}
