% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_immunarch_bycelltype.R
\name{prepare_immunarch_by_celltype}
\alias{prepare_immunarch_by_celltype}
\title{Prepare immunarch type of data across cell types}
\usage{
prepare_immunarch_by_celltype(
  immunarch_data_tra = NULL,
  immunarch_data_trb = NULL,
  input_string = NULL,
  cell_meta,
  filter_chain = "tra",
  min_cells = 1000,
  ncores = parallel::detectCores(),
  verbose = T
)
}
\arguments{
\item{immunarch_data_tra}{immunarch format for tra}

\item{immunarch_data_trb}{immunarch format for trb}

\item{input_string}{string to define tree structure}

\item{cell_meta}{cell metadata}

\item{filter_chain}{specify chain to be filtered, can be one of 'tra' or 'trb' or 'tra+trb'.}

\item{min_cells}{minimum number of cells per cell type}

\item{ncores}{number of cores}

\item{verbose}{Default is T}
}
\description{
Prepare immunarch type of data across cell types
}
