% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUnqKmers.R
\name{split_to_unq_kmers}
\alias{split_to_unq_kmers}
\title{Count unique kmers}
\usage{
split_to_unq_kmers(.data, .k, ncores = parallel::detectCores())
}
\arguments{
\item{.data}{list of sequences}

\item{.k}{Integer. Size of kmers.}

\item{ncores}{number of cores}
}
\description{
Count unique kmers
}
