% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treecor_GO.R
\name{treecor_GO}
\alias{treecor_GO}
\title{Perform GO analysis}
\usage{
treecor_GO(
  hierarchy_list,
  deg_list,
  background_gene,
  ontology_name = "BP",
  annotation_function = "annFUN.org",
  mapping = "org.Hs.eg.db",
  algorithm = "classic",
  statistic = "fisher",
  min_DEGs = 50,
  min_annotated_genes = 10,
  fdr_cutoff = 0.05,
  fold_change_cutoff = 1.5,
  ncores = parallel::detectCores(),
  verbose = T
)
}
\arguments{
\item{hierarchy_list}{hierarchy list}

\item{deg_list}{differential expressed genes, can be obtained by running `treecor_deg_iwas` or `treecor_deg`}

\item{background_gene}{specify a set of background genes}

\item{annotation_function}{specify an annotation function. Default is 'annFUN.org'.}

\item{mapping}{specify mapping database, default is 'org.Hs.eg.db'.}

\item{algorithm}{algorithm to be used in runTest}

\item{statistic}{statistic to be used in runTest}

\item{min_DEGs}{proceed with GO analysis with at least `min_DEGs` (default = 50)}

\item{min_annotated_genes}{retain GO terms with at least `min_annotated_genes` (default = 10)}

\item{fdr_cutoff}{FDR cutoff (default = 0.05)}

\item{fold_change_cutoff}{Fold change cutoff (default = 1.5)}

\item{ncores}{multiple cores}

\item{verbose}{show progress}

\item{ontology}{specify ontology to be used in GO analysis. Default is 'BP' (biological process).}
}
\description{
Perform GO analysis
}
