% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treecor_kmer.R
\name{treecor_kmer}
\alias{treecor_kmer}
\title{treecor_kmer}
\usage{
treecor_kmer(
  immunarch_list,
  chain,
  k = 5,
  unq_kmer_per_seq = T,
  remove_residual = T,
  start_end_pos = c(4, -4),
  norm = T,
  hierarchy_list,
  sample_meta,
  response_variable,
  separate = T,
  differential_method = "limma",
  weight = NULL,
  formula = NULL,
  coef = 2,
  fdr_cutoff = 0.05,
  ncores = parallel::detectCores(),
  save_as_csv = T,
  verbose = T
)
}
\arguments{
\item{immunarch_list}{A list of immune profiling info in immunarch format. Default is NULL. Users can provide their processed pseudobulk list (e.g. after covariate adjustment) via this parameter. Note that the names of list shall be matched with \code{`id`} extracted from \code{`hierarchy_list`}.}

\item{chain}{specify which tcr chain: 'tra' or 'trb'.}

\item{k}{kmer}

\item{unq_kmer_per_seq}{an indicator to specify whether to count only unique kmer}

\item{remove_residual}{specify whether remove residuals and keep only contact region}

\item{start_end_pos}{if remove residual is true, then specify starting and ending positions. Default is c(4,-4), which indicates to start at 4th amino acid and ends at 4th from the end.}

\item{norm}{normalize to frequency}

\item{hierarchy_list}{a hierarchy list by running \code{'extract_hrchy_string()'} or \code{'extract_hrchy_seurat()'} function.}

\item{sample_meta}{Sample-level metadata, where each row is a sample. Must contain 'sample' column and additional variables to be used in the analysis, such as covariates or outcomes of interest.}

\item{response_variable}{A vector of response variables.}

\item{separate}{A TRUE (default) or FALSE indicator, specifying how to evaluate multivariate outcomes.
\itemize{
\item TRUE: evaluate multivariate phenotype separately (it is equivalent to run this pipeline for each univariate phenotype).
\item FALSE: evaluate multivariate phenotype jointly.
}}

\item{differential_method}{Specify which type of differential methods: limma (default) or gam}

\item{weight}{A weight matrix to combine multivariate phenotype. The dimension should be number_phenotype * 1 If none is provided, then PC1 will be used as a joint univariate phenotype.}

\item{formula}{An object of class 'formula': a symbolic description of adjustment formula (i.e. only includes covariates other than response variable(s))}

\item{coef}{A column number or column name specifying which coefficient to be extracted (by default: 2).}

\item{fdr_cutoff}{Cutoff value for FDR. Only genes with lower FDR are listed. Default is 0.05.}

\item{ncores}{Number of cores to be used. If ncores > 1, it will be implemented in a parallel mode.}

\item{save_as_csv}{An indicator to save identified DEGs in csv files. DEGs for each cell cluster is saved as 'responsevariable_celltype_DEG.csv' and a summary file of combining DEGs from all cell clusters is saved as 'responsevariable_combinedDEG.csv'.}

\item{verbose}{Show progress}
}
\value{
A list of four elements:
\itemize{
\item diff.summary: A summary table of number of differential features for each tree node.
\item diff.ls: A comprehensive list of outcome(s)-associated feature set for each tree node. Use \code{`result$dge.ls$response_variable[[celltype]]`} to extract DEGs for a specific cell type
\item feature.ls: A list of processed feature set for each cell type
}
}
\description{
treecor_kmer
}
\author{
Boyang Zhang <bzhang34@jhu.edu>, Hongkai Ji
}
